requirements.txt
# Core dependencies
langchain-core>=0.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
langgraph-cli[inmem]>=0.4.4

# Document processing
python-docx>=1.1.0
pypandoc>=1.13

# LLM and API
openai>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Development and testing
langgraph-sdk>=0.1.0
